# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DSCR Loan Automation Platform API
  description: |
    Comprehensive API for the DSCR (Debt Service Coverage Ratio) Loan Automation Platform.

    This platform automates the full lifecycle of DSCR refinance loans from lead intake
    through funding, with native integration to ICE Mortgage Technology Encompass LOS.

    ## Authentication
    All endpoints require Bearer token authentication via OAuth 2.0.

    ## Rate Limiting
    API calls are limited to 1000 requests per minute per client.

    ## Versioning
    The API is versioned via URL path (e.g., /api/v1/).
  version: 1.0.0
  contact:
    name: DSCR Platform Team
    email: support@dscrplatform.com
servers:
  - url: https://api.dscrplatform.com/api/v1
    description: Production
  - url: https://staging-api.dscrplatform.com/api/v1
    description: Staging
...
paths:
...
  /applications/{applicationId}/avm:
    post:
      tags: [Valuation]
      summary: Order AVM
      operationId: orderAVM
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '202':
          description: AVM order accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AVMReport'

    get:
      tags: [Valuation]
      summary: Get AVM reports
      operationId: getAVMReports
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: AVM reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AVMReport'
...
components:
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Application ID
...
  schemas:
...
    Money:
      type: integer
      description: Amount in cents
...
    AVMReport:
      type: object
      properties:
        id:
          type: string
        estimatedValue:
          $ref: '#/components/schemas/Money'
        confidenceLevel:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        confidenceScore:
          type: integer
        valueLow:
          $ref: '#/components/schemas/Money'
        valueHigh:
          $ref: '#/components/schemas/Money'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []